Struct katex::opts::Opts

#[non_exhaustive]
pub struct Opts { /* private fields */ }

Options to be passed to KaTeX.

Read https://katex.org/docs/options.html for more information.

Implementations

impl Opts

pub fn builder() -> OptsBuilder

Return OptsBuilder.

pub fn set_display_mode(&mut self, flag: bool)

Set whether to render the math in the display mode.

pub fn set_output_type(&mut self, output_type: OutputType)

Set KaTeX output type.

pub fn set_leqno(&mut self, flag: bool)

Set whether to have \tags rendered on the left instead of the right.

pub fn set_fleqn(&mut self, flag: bool)

Set whether to make display math flush left.

pub fn set_throw_on_error(&mut self, flag: bool)

Set whether to let KaTeX throw a ParseError for invalid LaTeX.

pub fn set_error_color(&mut self, color: String)

Set the color used for invalid LaTeX.

pub fn add_macro(&mut self, entry_name: String, entry_data: String)

Add a custom macro. Read https://katex.org/docs/options.html for more information.

pub fn set_min_rule_thickness(&mut self, value: f64)

Set the minimum thickness, in ems. Read https://katex.org/docs/options.html for more information.

pub fn set_max_size(&mut self, value: Option<f64>)

Set the max size for user-specified sizes. If set to None, users can make elements and spaces arbitrarily large. Read https://katex.org/docs/options.html for more information.

pub fn set_max_expand(&mut self, value: Option<i32>)

Set the limit for the number of macro expansions. If set to None, the macro expander will try to fully expand as in LaTeX. Read https://katex.org/docs/options.html for more information.

pub fn set_trust(&mut self, flag: bool)

Set whether to trust users’ input. Read https://katex.org/docs/options.html for more information.