Struct katex::opts::OptsBuilder

pub struct OptsBuilder { /* private fields */ }

Builder for Opts.

Implementations

impl OptsBuilder

pub fn display_mode<VALUE: Into<bool>>(&mut self, value: VALUE) -> &mut Self

Whether to render the math in the display mode.

pub fn output_type<VALUE: Into<OutputType>>( &mut self, value: VALUE ) -> &mut Self

KaTeX output type.

pub fn leqno<VALUE: Into<bool>>(&mut self, value: VALUE) -> &mut Self

Whether to have \tags rendered on the left instead of the right.

pub fn fleqn<VALUE: Into<bool>>(&mut self, value: VALUE) -> &mut Self

Whether to make display math flush left.

pub fn throw_on_error<VALUE: Into<bool>>(&mut self, value: VALUE) -> &mut Self

Whether to let KaTeX throw a ParseError for invalid LaTeX.

pub fn error_color<VALUE: Into<String>>(&mut self, value: VALUE) -> &mut Self

Color used for invalid LaTeX.

pub fn macros<VALUE: Into<HashMap<String, String>>>( &mut self, value: VALUE ) -> &mut Self

Collection of custom macros. Read https://katex.org/docs/options.html for more information.

pub fn min_rule_thickness<VALUE: Into<f64>>( &mut self, value: VALUE ) -> &mut Self

Specifies a minimum thickness, in ems. Read https://katex.org/docs/options.html for more information.

pub fn max_size<VALUE: Into<Option<f64>>>(&mut self, value: VALUE) -> &mut Self

Max size for user-specified sizes. If set to None, users can make elements and spaces arbitrarily large. Read https://katex.org/docs/options.html for more information.

pub fn max_expand<VALUE: Into<Option<i32>>>( &mut self, value: VALUE ) -> &mut Self

Limit the number of macro expansions to the specified number. If set to None, the macro expander will try to fully expand as in LaTeX. Read https://katex.org/docs/options.html for more information.

pub fn trust<VALUE: Into<bool>>(&mut self, value: VALUE) -> &mut Self

Whether to trust users’ input. Read https://katex.org/docs/options.html for more information.

pub fn build(&self) -> Result<Opts, OptsBuilderError>

Builds a new Opts.

Errors

If a required field has not been initialized.

impl OptsBuilder

pub fn add_macro(self, entry_name: String, entry_data: String) -> Self

Add an entry to macros.

Examples

let opts = katex::Opts::builder()
    .add_macro(r#"\RR"#.to_owned(), r#"\mathbb{R}"#.to_owned())
    .build()
    .unwrap();
let html = katex::render_with_opts(r#"\RR"#, &opts).unwrap();